Testing: AdminController commented

+++ src/main/java/com/ingeacev/reto3/dbo/ClientDbo.java
@@ -19,7 +19,7 @@ public class ClientDbo {
 
     private String password;
 
-    private int age;
+    private byte age;
 
     private String name;
 }
